Time: 107 ms (27.87%), Space: 64.9 MB (24.32%) - LeetHub

+++ 0287-find-the-duplicate-number/0287-find-the-duplicate-number.cpp
@@ -4,7 +4,7 @@ public:
         int ans;
         sort(nums.begin(), nums.end());
 
-        for (int i=0; i<nums.size(); i++) {
+        for (size_t i=0; i<nums.size(); i++) {
             if (nums[i] == nums[i+1]) {
                 ans = nums[i];
                 break;
